pc/requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
PyYAML>=6.0.1
ultralytics>=8.3.0
